tests/wpt/web-platform-tests/css/css-ui/parsing/outline-valid-optional.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS UI Level 3: parsing outline with valid values</title>
<link rel="author" title="Eric Willigers" href="mailto:ericwilligers@chromium.org">
<link rel="help" href="https://drafts.csswg.org/css-ui-3/#outline">
<link rel="help" href="https://drafts.csswg.org/cssom/#serializing-css-values">
<meta name="assert" content="outline supports the full grammar '<outline-color> || <outline> || <outline>'.">
<meta name="assert" content="outline serializes in canonical order, with shortest possible serialization.">
<meta name="flags" content="may">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/parsing-testcommon.js"></script>
</head>
<body>
<script>
// Conformant UAs may ignore the invert value on platforms that do not support color inversion of the pixels on the screen.
test_valid_value("outline", "invert");
test_valid_value("outline", "invert dotted 1px");
</script>
</body>
</html>
